Puts DefaultProxyFilter into its own file
